08-spring-orm-jpa/bookkeeper/src/main/java/ru/dartilla/bookkeeper/repositores/CommentRepositoryJpa.java
package ru.dartilla.bookkeeper.repositores;

import org.springframework.stereotype.Repository;
import ru.dartilla.bookkeeper.domain.Comment;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import javax.persistence.TypedQuery;
import java.util.List;
import java.util.Optional;

@Repository
public class CommentRepositoryJpa implements CommentRepository {

    @PersistenceContext
    private EntityManager em;

    @Override
    public Optional<Comment> findById(Long id) {
        return Optional.ofNullable(em.find(Comment.class, id));
    }

    @Override
    public void save(Comment comment) {
        em.persist(comment);
    }

    @Override
    public List<Comment> findByScript(Long scriptId) {
        TypedQuery<Comment> query = em.createQuery("select c from Comment c where c.script.id = :scriptId",
                Comment.class);
        query.setParameter("scriptId", scriptId);
        return query.getResultList();
    }
}
